--- src/main/java/http/View.java
@@ -1,8 +1,0 @@
-package http;
-
-import http.request.HttpRequest;
-import http.response.HttpResponse;
-
-public interface View {
-    void render(HttpRequest request, HttpResponse response) throws Exception;
-}
