+++ src/main/java/view/RedirectView.java
@@ -1,7 +1,9 @@
-package http;
+package view;
 
 import http.request.HttpRequest;
 import http.response.HttpResponse;
+
+import java.util.Map;
 
 public class RedirectView implements View {
     private String path;
@@ -11,7 +13,7 @@ public class RedirectView implements View {
     }
 
     @Override
-    public void render( HttpRequest request, HttpResponse response) {
+    public void render(Map<String, ?> model, HttpRequest request, HttpResponse response) {
         response.redirect(path);
     }
 }
