+++ src/main/java/view/TemplateView.java
@@ -1,7 +1,9 @@
-package http;
+package view;
 
 import http.request.HttpRequest;
 import http.response.HttpResponse;
+
+import java.util.Map;
 
 public class TemplateView implements View {
     private String path;
@@ -11,7 +13,7 @@ public class TemplateView implements View {
     }
 
     @Override
-    public void render(HttpRequest request, HttpResponse response) throws Exception {
+    public void render(Map<String, ?> model, HttpRequest request, HttpResponse response) throws Exception {
         response.forward(request.getPath() + path);
     }
 }
