+++ src/main/java/view/View.java
@@ -1,0 +1,10 @@
+package view;
+
+import http.request.HttpRequest;
+import http.response.HttpResponse;
+
+import java.util.Map;
+
+public interface View {
+    void render(Map<String, ?> model, HttpRequest request, HttpResponse response) throws Exception;
+}
